+++ src/test/java/org/opentripplanner/standalone/config/RouterConfigTest.java
@@ -1,53 +1,34 @@
 package org.opentripplanner.standalone.config;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.fail;
-import static org.opentripplanner.standalone.config.framework.json.JsonSupport.jsonNodeForTest;
-import static org.opentripplanner.standalone.config.framework.json.JsonSupport.jsonNodeFromResource;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.opentripplanner.standalone.config.framework.json.JsonSupport.newNodeAdapterForTest;
 
+import java.time.Duration;
 import org.junit.jupiter.api.Test;
-import org.opentripplanner.framework.application.OtpAppException;
-import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
 
 class RouterConfigTest {
 
-  private static final String SOURCE = "RouterConfigTest";
-
-  /**
-   * Test that the router-config.json example used in documentation is valid.
-   */
   @Test
-  void validateExample() {
-    var node = jsonNodeFromResource("standalone/config/router-config.json");
-
-    // Setup so we get access to the NodeAdapter
-    var a = new NodeAdapter(node, SOURCE);
-    var c = new RouterConfig(a, false);
-
-    // Test for unused parameters
-    var buf = new StringBuilder();
-    a.logAllWarnings(m -> buf.append("\n").append(m));
-    if (!buf.isEmpty()) {
-      fail(buf.toString());
-    }
+  void defaultMaxSearchWindowIs24Hours() {
+    validateMaxSearchWindow("", Duration.ofHours(24));
   }
 
   @Test
-  void testSemanticValidation() {
-    // apiProcessingTimeout must be greater then streetRoutingTimeout
-    var root = createNodeAdaptor(
+  void maxSearchWindowIsOverriddenInRouterConfig() {
+    validateMaxSearchWindow(
       """
       {
-        server: { apiProcessingTimeout : "1s" },
-        routingDefaults: { streetRoutingTimeout: "17s" }
+        "transit": {
+          "maxSearchWindow": "48h"
+        }
       }
-      """
+      """,
+      Duration.ofHours(48)
     );
-    //
-    assertThrows(OtpAppException.class, () -> new RouterConfig(root, false));
   }
 
-  private static NodeAdapter createNodeAdaptor(String jsonText) {
-    return new NodeAdapter(jsonNodeForTest(jsonText), "Test");
+  private void validateMaxSearchWindow(String configuration, Duration expectedDuration) {
+    RouterConfig routerConfig = new RouterConfig(newNodeAdapterForTest(configuration), false);
+    assertEquals(expectedDuration, routerConfig.routingRequestDefaults().maxSearchWindow());
   }
 }
